Peers Pagination logic extracted and moved to components
